Move FormController to javarosa package
