+++ collect_app/src/androidTest/java/org/odk/collect/android/support/CollectHelpers.java
@@ -6,7 +6,7 @@ import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.injection.config.AppDependencyComponent;
 import org.odk.collect.android.injection.config.AppDependencyModule;
 import org.odk.collect.android.injection.config.DaggerAppDependencyComponent;
-import org.odk.collect.android.logic.FormController;
+import org.odk.collect.android.formentry.javarosawrapper.FormController;
 
 public final class CollectHelpers {
