+++ collect_app/src/main/java/org/odk/collect/android/activities/ViewOnlyFormHierarchyActivity.java
@@ -19,7 +19,7 @@ import android.widget.Button;
 
 import org.javarosa.core.model.FormIndex;
 import org.odk.collect.android.R;
-import org.odk.collect.android.logic.FormController;
+import org.odk.collect.android.formentry.javarosawrapper.FormController;
 
 /**
  * Displays the structure of a form along with the answers for the current instance. Disables all
