+++ collect_app/src/main/java/org/odk/collect/android/formentry/FormControllerProvider.java
@@ -1,6 +1,6 @@
 package org.odk.collect.android.formentry;
 
-import org.odk.collect.android.logic.FormController;
+import org.odk.collect.android.formentry.javarosawrapper.FormController;
 
 public interface FormControllerProvider {
