+++ collect_app/src/main/java/org/odk/collect/android/formentry/backgroundlocation/BackgroundLocationHelper.java
@@ -5,6 +5,7 @@ import android.location.Location;
 import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.formentry.audit.AuditConfig;
 import org.odk.collect.android.formentry.audit.AuditEvent;
+import org.odk.collect.android.formentry.javarosawrapper.FormController;
 import org.odk.collect.android.preferences.GeneralSharedPreferences;
 import org.odk.collect.android.utilities.PermissionUtils;
 import org.odk.collect.android.utilities.PlayServicesUtil;
@@ -16,7 +17,7 @@ import static org.odk.collect.android.preferences.GeneralKeys.KEY_BACKGROUND_LOC
  *
  * Ideally this would be replaced by more coherent abstractions in the future.
  *
- * The methods on the {@link org.odk.collect.android.logic.FormController} are wrapped here rather
+ * The methods on the {@link FormController} are wrapped here rather
  * than the form controller being passed in when constructing the {@link BackgroundLocationManager}
  * because the form controller isn't set until
  * {@link org.odk.collect.android.activities.FormEntryActivity}'s onCreate.
