+++ collect_app/src/main/java/org/odk/collect/android/formentry/saving/DiskFormSaver.java
@@ -3,7 +3,7 @@ package org.odk.collect.android.formentry.saving;
 import android.net.Uri;
 
 import org.odk.collect.android.analytics.Analytics;
-import org.odk.collect.android.logic.FormController;
+import org.odk.collect.android.formentry.javarosawrapper.FormController;
 import org.odk.collect.android.tasks.SaveFormToDisk;
 import org.odk.collect.android.tasks.SaveToDiskResult;
