+++ collect_app/src/main/java/org/odk/collect/android/utilities/FormEntryPromptUtils.java
@@ -27,7 +27,7 @@ import org.javarosa.core.model.data.MultipleItemsData;
 import org.javarosa.core.model.data.helper.Selection;
 import org.javarosa.form.api.FormEntryPrompt;
 import org.odk.collect.android.dao.ItemsetDao;
-import org.odk.collect.android.logic.FormController;
+import org.odk.collect.android.formentry.javarosawrapper.FormController;
 
 import java.math.BigDecimal;
 import java.text.DecimalFormat;
